[#140] refactor: 메인, 랜딩 페이지 위치변경
