+++ src/app/page.tsx
@@ -1,23 +1,14 @@
-import MainMapSection from '@/components/mainPage/MainMapSection';
-import MainFooter from '@/components/mainPage/MainFooter';
-import MainRecommendSection from '@/components/mainPage/MainRecommendSection';
-import MainStampSection from '@/components/mainPage/MainStampSection';
+import LandingPage from '@/components/landing/Landingpage';
+import React from 'react';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  title: '홈',
-  description: '홈 페이지 입니다.'
+  title: '서비스 소개',
+  description: '서비스 소개 페이지 입니다.'
 };
 
-const MainHomePage = () => {
-  return (
-    <main className="lg:mt-18 relative mt-20">
-      <MainMapSection />
-      <MainStampSection />
-      <MainRecommendSection />
-      <MainFooter />
-    </main>
-  );
+const page = () => {
+  return <LandingPage />;
 };
 
-export default MainHomePage;
+export default page;
